refactor: split app into modules and add bundler
