+++ vite.config.js
@@ -1,0 +1,12 @@
+import { defineConfig } from 'vite';
+
+export default defineConfig({
+  root: 'web',
+  build: {
+    outDir: '../dist',
+    emptyOutDir: true
+  },
+  define: {
+    DEBUG: false
+  }
+});
