+++ web/index.html
@@ -121,6 +121,6 @@
     </script>
 
     <!-- App principal -->
-    <script type="module" src="./app.js"></script>
+    <script type="module" src="./main.js"></script>
   </body>
 </html>
